Rename /client/typings to /client/types/
