+++ client/src/test/test-settings/testSettingsSlice.ts
@@ -1,5 +1,5 @@
 import { createSlice, PayloadAction } from "@reduxjs/toolkit";
-import { TestSettingsState } from "client/typings/reduxTypes";
+import { TestSettingsState } from "client/types/reduxTypes";
 
 const initialState: TestSettingsState = {
   testTime: "",
