+++ client/src/types/declarations/custom.d.ts
