+++ client/src/types/declarations/i18next.d.ts
