+++ client/src/utils/hooks.ts
@@ -1,5 +1,5 @@
 import { TypedUseSelectorHook, useDispatch, useSelector } from "react-redux";
-import { AppDispatch, RootState } from "client/typings/reduxTypes";
+import { AppDispatch, RootState } from "client/types/reduxTypes";
 
 // Use instead of plain `useDispatch`
 /* eslint-disable-next-line ban/ban */
