+++ client/src/views/admin/adminThunks.ts
@@ -6,7 +6,7 @@ import {
   postSettings,
 } from "client/services/settingsServices";
 import { Game } from "shared/typings/models/game";
-import { AppThunk } from "client/typings/reduxTypes";
+import { AppThunk } from "client/types/reduxTypes";
 import {
   submitUpdateHiddenAsync,
   submitGetSettingsAsync,
