+++ client/src/views/group/groupThunks.ts
@@ -5,7 +5,7 @@ import {
   postLeaveGroup,
   postCloseGroup,
 } from "client/services/groupServices";
-import { AppThunk } from "client/typings/reduxTypes";
+import { AppThunk } from "client/types/reduxTypes";
 import {
   submitLeaveGroupAsync,
   submitUpdateGroupAsync,
